migrate subscription and support
