+++ app/cmd/routes.go
@@ -113,10 +113,10 @@ func routes(r *web.Engine) *web.Engine {
 			private.Post("/api/posts/:number/comments", handlers.PostComment())
 			private.Post("/api/posts/:number/comments/:id", handlers.UpdateComment())
 			private.Put("/api/v1/posts/:number/status", apiv1.SetResponse())
-			private.Post("/api/posts/:number/support", handlers.AddSupporter())
-			private.Post("/api/posts/:number/unsupport", handlers.RemoveSupporter())
-			private.Post("/api/posts/:number/subscribe", handlers.Subscribe())
-			private.Post("/api/posts/:number/unsubscribe", handlers.Unsubscribe())
+			private.Post("/api/v1/posts/:number/support", apiv1.AddSupporter())
+			private.Delete("/api/v1/posts/:number/support", apiv1.RemoveSupporter())
+			private.Post("/api/v1/posts/:number/subscription", apiv1.Subscribe())
+			private.Delete("/api/v1/posts/:number/subscription", apiv1.Unsubscribe())
 			private.Post("/api/posts/:number/tags/:slug", handlers.AssignTag())
 			private.Delete("/api/posts/:number/tags/:slug", handlers.UnassignTag())
 			private.Delete("/api/user", handlers.DeleteUser())
