+++ app/handlers/apiv1/post_test.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/getfider/fider/app/handlers"
 	"github.com/getfider/fider/app/handlers/apiv1"
 	"github.com/getfider/fider/app/models"
 	. "github.com/getfider/fider/app/pkg/assert"
@@ -141,7 +140,7 @@ func TestSetResponseHandler(t *testing.T) {
 		OnTenant(mock.DemoTenant).
 		AsUser(mock.JonSnow).
 		AddParam("number", post.ID).
-		ExecutePost(handlers.SetResponse(), fmt.Sprintf(`{ "status": %d, "text": "Done!" }`, models.PostCompleted))
+		ExecutePost(apiv1.SetResponse(), fmt.Sprintf(`{ "status": %d, "text": "Done!" }`, models.PostCompleted))
 
 	post, _ = services.Posts.GetByNumber(post.Number)
 
@@ -164,7 +163,7 @@ func TestSetResponseHandler_Unauthorized(t *testing.T) {
 		OnTenant(mock.DemoTenant).
 		AsUser(mock.AryaStark).
 		AddParam("number", post.ID).
-		ExecutePost(handlers.SetResponse(), fmt.Sprintf(`{ "status": %d, "text": "Done!" }`, models.PostCompleted))
+		ExecutePost(apiv1.SetResponse(), fmt.Sprintf(`{ "status": %d, "text": "Done!" }`, models.PostCompleted))
 
 	post, _ = services.Posts.GetByNumber(post.Number)
 
@@ -186,7 +185,7 @@ func TestSetResponseHandler_Duplicate(t *testing.T) {
 		OnTenant(mock.DemoTenant).
 		AsUser(mock.JonSnow).
 		AddParam("number", post1.ID).
-		ExecutePost(handlers.SetResponse(), body)
+		ExecutePost(apiv1.SetResponse(), body)
 	Expect(code).Equals(http.StatusOK)
 
 	post1, _ = services.Posts.GetByNumber(post1.Number)
@@ -209,7 +208,7 @@ func TestSetResponseHandler_Duplicate_NotFound(t *testing.T) {
 		OnTenant(mock.DemoTenant).
 		AsUser(mock.JonSnow).
 		AddParam("number", post1.ID).
-		ExecutePost(handlers.SetResponse(), body)
+		ExecutePost(apiv1.SetResponse(), body)
 
 	Expect(code).Equals(http.StatusBadRequest)
 }
@@ -227,7 +226,65 @@ func TestSetResponseHandler_Duplicate_Itself(t *testing.T) {
 		OnTenant(mock.DemoTenant).
 		AsUser(mock.JonSnow).
 		AddParam("number", post.ID).
-		ExecutePost(handlers.SetResponse(), body)
+		ExecutePost(apiv1.SetResponse(), body)
 
 	Expect(code).Equals(http.StatusBadRequest)
+}
+
+func TestAddSupporterHandler(t *testing.T) {
+	RegisterT(t)
+
+	server, services := mock.NewServer()
+	services.SetCurrentTenant(mock.DemoTenant)
+	services.SetCurrentUser(mock.JonSnow)
+	first, _ := services.Posts.Add("The Post #1", "The Description #1")
+	second, _ := services.Posts.Add("The Post #2", "The Description #2")
+
+	code, _ := server.
+		OnTenant(mock.DemoTenant).
+		AsUser(mock.AryaStark).
+		AddParam("number", second.Number).
+		Execute(apiv1.AddSupporter())
+
+	first, _ = services.Posts.GetByNumber(1)
+	second, _ = services.Posts.GetByNumber(2)
+
+	Expect(code).Equals(http.StatusOK)
+	Expect(first.TotalSupporters).Equals(0)
+	Expect(second.TotalSupporters).Equals(1)
+}
+
+func TestAddSupporterHandler_InvalidPost(t *testing.T) {
+	RegisterT(t)
+
+	server, _ := mock.NewServer()
+	code, _ := server.
+		OnTenant(mock.DemoTenant).
+		AsUser(mock.AryaStark).
+		AddParam("number", 999).
+		Execute(apiv1.AddSupporter())
+
+	Expect(code).Equals(http.StatusNotFound)
+}
+
+func TestRemoveSupporterHandler(t *testing.T) {
+	RegisterT(t)
+
+	server, services := mock.NewServer()
+	services.SetCurrentTenant(mock.DemoTenant)
+	services.SetCurrentUser(mock.JonSnow)
+	post, _ := services.Posts.Add("The Post #1", "The Description #1")
+	services.Posts.AddSupporter(post, mock.JonSnow)
+	services.Posts.AddSupporter(post, mock.AryaStark)
+
+	code, _ := server.
+		OnTenant(mock.DemoTenant).
+		AsUser(mock.AryaStark).
+		AddParam("number", post.ID).
+		Execute(apiv1.RemoveSupporter())
+
+	post, _ = services.Posts.GetByNumber(post.Number)
+
+	Expect(code).Equals(http.StatusOK)
+	Expect(post.TotalSupporters).Equals(1)
 }
