+++ app/handlers/post.go
@@ -150,53 +150,6 @@ func UpdateComment() web.HandlerFunc {
 	}
 }
 
-// AddSupporter adds current user to given post list of supporters
-func AddSupporter() web.HandlerFunc {
-	return func(c web.Context) error {
-		return addOrRemove(c, c.Services().Posts.AddSupporter)
-	}
-}
-
-// RemoveSupporter removes current user from given post list of supporters
-func RemoveSupporter() web.HandlerFunc {
-	return func(c web.Context) error {
-		return addOrRemove(c, c.Services().Posts.RemoveSupporter)
-	}
-}
-
-// Subscribe adds current user to list of subscribers of given post
-func Subscribe() web.HandlerFunc {
-	return func(c web.Context) error {
-		return addOrRemove(c, c.Services().Posts.AddSubscriber)
-	}
-}
-
-// Unsubscribe removes current user from list of subscribers of given post
-func Unsubscribe() web.HandlerFunc {
-	return func(c web.Context) error {
-		return addOrRemove(c, c.Services().Posts.RemoveSubscriber)
-	}
-}
-
-func addOrRemove(c web.Context, addOrRemove func(post *models.Post, user *models.User) error) error {
-	number, err := c.ParamAsInt("number")
-	if err != nil {
-		return c.Failure(err)
-	}
-
-	post, err := c.Services().Posts.GetByNumber(number)
-	if err != nil {
-		return c.Failure(err)
-	}
-
-	err = addOrRemove(post, c.User())
-	if err != nil {
-		return c.Failure(err)
-	}
-
-	return c.Ok(web.Map{})
-}
-
 // ExportPostsToCSV returns a CSV with all posts
 func ExportPostsToCSV() web.HandlerFunc {
 	return func(c web.Context) error {
